moved CuPyCPUCompressor to the test suite
